Added rofi configuration
Also edited the scripts for the power menu and the power profiles menu so
that it looks good. The icons were clipping into the words previously.

The rofi configuration is mainly just to set the default DPI for rofi in
the config.rasi file. The other .rasi files are the theme files for the
i3 script and are provided by EndeavourOS. I did a little editing to the
theme files for the power menu and the power profiles menu. Also changed
the font size for the rofi dmenu so that it's not so small as for some
reason the dmenu doesn't follow the DPI setting set in the rofi
config.rasi file.