Used the "lazy" variable for firenvim instead
The load function in lazy.nvim somehow doesn't load firenvim even when
it is forced with the "wait = true". As such, I have to change the
configuration to use the "lazy" variable instead to not load firenvim
when using Neovim in the terminal but still have firenvim load when it
is started in the browser.