Updated the firenvim configuration
Now that @folke has changed the functionality of the cond keyword, the
only sane way to lazy load firenvim is to use the lazy keyword to load
firenvim when needed.

Also removed the lazy load function to load firenvim as it is no longer
needed. Firenvim will be loaded when the firenvim#install function is
called.