Used vim keybindings for window management
Switching the focused window and moving the focused window now use vim
hjkl bindings instead of the default, where h is for left, j is for
down, k is for up and l is for right.

The keybinding for the horizontal split has been changed to $mod + b
from $mod + h and the keybinding for locking the screen has been changed
to $mod + o from $mod + l, since $mod + h and $mod + l are now used for
changing the current focused window.