Added a section about literate programming
With that section, configured org mode to load more languages such as
Python so that I can use org mode to create Jupyter notebooks with
Jupytext.

Also stopped org mode from asking me to confirm if I want to evaluate
the buffer because that would get old very quickly.

This is all in preparation for the CY1308 Physics and CY1601 Math I
modules, which will use Jupyter notebooks.