Changed the theme to bluloco dark
Hell yeahhhhhhhh. The bluloco themes have finally been merged into the
doom-themes repository, which is why I'm using them now. So glad that
I'm finally able to use the theme that I use for Neovim in Emacs now.

With this change, the uwu theme is still installed, but it is no
longer loaded since bluloco dark will be the default theme that is
loaded on startup from now on. I'm just leaving the uwu theme in for
legacy purposes, since it's a great high contrast theme that saw me
through the period without the bluloco dark theme.

Also, I uninstalled LSP treemacs as I have no use for it. I'm mostly
going to do development work on Neovim instead of Emacs since Neovim
is just much faster and more familiar to me, and everything in Neovim
works exactly the way I want it to work. I'm mostly going to just use
Emacs for creating notes with org-mode and LaTeX as Emacs is really
the best for that.