Use gummy instead of brightnessctl
Brightnessctl hasn't been updated in 1.5 years. It works well for the
most part, but I don't really like using unmaintained packages. The
maintainer stated that he will try being a proper maintainer in the
second half of August, but that has since passed and it's now September,
with no new commits to the repository.

One other big thing is that gummy does screen temperature as well, which
is something I have needed for the longest time. A lot of redshifting
applications on Linux have basically been abandoned, like Redshift,
mostly because GNOME and KDE both provide redshifters, usually called
night light, as a built-in application for their desktop environments so
there's not much of a need for other alternatives if you use either
GNOME or KDE. But I don't use either so I need a maintained
alternative for this feature.

Another thing that gummy does is that it allows you to set the pixel
brightness as well as the backlight brightness, which gives you a lot of
control over how you want the screen to look, as having a high backlight
brightness usually makes the screen look bluish when it's supposed to be
black, so setting the backlight brightness to a low value for blacker
blacks and less blue light is great. All this is assuming you have an LCD
screen on your monitor, as OLED screens don't have a backlight, so you
can only set the pixel brightness on OLED screens.

I'm glad to have stumbled across Gammy in a Reddit thread, which led me
to find gummy, which is the linux-only successor to Gammy as Gammy used
to be cross-platform and worked on Linux, Windows and FreeBSD, but it is no
longer being maintained.