Rewrote the lsp configuration for lua ls
This is so that it is more flexible and supports other configuration
files in the .config directory, like wezterm and awesomewm configuration
files.

Before this, lua ls will refuse to initialise for other
configuration files as the git repository is in the home directory. With
these changes, lua ls will initialise properly and can be used to help
in the configuration of these other programs.

It is also more clear what the lua ls settings are instead of being hidden
behind lsp-zero's nvim_lua_ls function.