Cleaned up the Neovim config
Also added null-ls with mason-null-ls to configure additional formatters
and linters, such as eslint and prettier as tsserver is just not enough
for working with svelte files.

Also added mini.files as the file explorer, which works like lf and
manipulates the directory using text editing, which is excellent. This
means that the keybinding for netrw has been removed.

Also added gitsigns so I know what has changed in a file.

Mason configuration has been pulled out of the LSP configuration so that
can be lazy loaded outside of the entire LSP configuration.

The close buffer plugin has been changed from vim-bbye to
mini.bufremove, which is actually maintained unlike vim-bbye.

Renamed a lot of plugin files to be more generic so swapping to a
different plugin that does the same thing isn't as painful.