Updated the configuration for whitespace mode
Whitespace mode now shows tabs, empty lines at the beginning or the
end of the buffer, and the missing newline character at the end of a
buffer.

Whitespace mode also shows the tab character now. Updated all the
characters shown to match with the Neovim listchars.

Also changed all mentions of hard spaces to non-breaking spaces to be
easier to understand.

Added a key bind to create a file.