Show the fill column at 80 characters
Also no longer show the empty lines at the beginning or at the end of a buffer
as it was pretty distracting and didn't do much for me. The auto cleanup from
whitespace mode also made it pretty useless.

Also stopped showing the missing newline character as it was also more
distracting that anything else.