Removed the key bind to search for a file using fzf
fzf just doesn't really work with the Emacs workflow as fzf is initialised in
the current directory that Emacs is in, which makes it pretty useless in
searching for files, unlike Neovim, which initialises fzf in the directory that
Neovim is started in instead of the current directory that Neovim is in.

Removed fzf as well as removed isync in the external programs as I have decided
to not use email in Emacs.