Added more features to the lf configuration
Added functions to copy the absolute path, directory path, file name
and the file name without the file extension to the clipboard. The
key bindings that these functions are bound to are the same as the Yazi
configuration. This requires xclip, but that's fine since I'm already
using xclip with Neovim.

Unmap the e, c and w keys.

Use Ctrl + s to open a shell instead of using w.

Use o to open a file with the default editor instead of using e.

The rename that removes the name of the file has been changed to cw
instead of just c.

ee is now used to extract a file instead of oa.

ec is now used to edit the lf configuration file instead of oe.

Use the function keyword for the lfcd command.