Added the Yazi config
Most of it is just mirroring the lf config.

Also added the ya command so that Yazi can change the shell directory
after exiting, just like the lfcd command that I already have.

Yazi is awesome, just so many things built-in, unlike lf which is super
barebones and needs a lot of external programs and shell scripts to make
it usable.