Add the theme config for yazi
Removed the init.lua file as removing the borders can be done by just
setting the border symbol to a space in the theme.toml file.

The theme configuration is mostly based on the lf configuration, but
with some additional files and coloured icons.

Also, changed some of the lf colours, particularly for archives and
images as they all look very similar, so the lf configuration uses
yazi's default colours for archives and images instead.