Fixed the yazi theme
Set the hover colour to be the same as the directory colour. The
previous one using a cyan colour was impossible to read. I wish it can
just be like lf and just take the colour of the item it is hovering and
fill the background with it, while also inverting the colour of the
words.

Set the mode colours to match the Neovim ones.

Also set the theme for the whichkey thing to match the Neovim one as well.

Set the colours for the help window, which were absolutely terrible and
made the help window absolutely useless due to it being impossible to
read. The cursor was extremely bright while the words were just way too
dark. The colours were set in accordance with the which key theme in
Neovim and also it also uses the selection colour in Neovim to show
where the cursor is hovering.

Added the muted white colour as a fallback for files as well.