Use the script for adjusting brightness
This is so that there is a notification when the brightness is being
adjusted.

The reason for using the gummyconf.json file instead of using the gummy
status command is because the gummy status command is quite slow to
update, which results in the brightness level not being accurate.

Also removed the icon when the volume notification is shown.