Changed the default search engine to DuckDuckGo
Brave with its AI nonsense is really getting on my nerves, and the AI
answers are inaccurate as well, which is just terrible for a search
engine having it as the first result.

DuckDuckGo doesn't seem to have the censorship issues that people has
mentioned, they probably rolled that back after all the backlash they
got.