Add back TabNine autocompletion as it works offline
Add Supermaven as an additional source to Codeium. Codeium still works
better, but Supermaven is extremely fast. Both require an internet
connection though.

Change cmp-path to cmp-async-path as it works without blocking Neovim which
makes it much faster and more responsive.

Also fixed the formatting and rearranged some of the completion sources
in the lsp.lua file.

Also updated the packages used for Neovim.