Removed lsp-zero entirely
My configuration was super customised anyways, and lsp-zero wasn't
really doing much, so now it is removed entirely.

Also reformatted the file and changed the diagnostic icons to a constant
table instead of a function.

Also use the dot notation instead of the square brackets.

Updated the README to reflect the changes.