Migrate all formatters from null-ls
The previous command for stylua didn't work to format the file with the
configuration present in the repository, which broke the language server
formatting. Now that it is fixed, I only need to use null-ls for
gitsigns and nothing else.

The formatter and linter configuration is now works across all editors.