Switch back to kitty
I think it's probably because Wezterm hasn't been updated in like almost
a year or so, but I having a heck ton of issues on Wezterm that I simply
don't face on Kitty.

Some examples are like the cursor flickering in Neogit windows when
typing out the commit message and centering the screen in one split pane
in Neovim somehow causes half of the other split pane to scroll along as
well, making the other split pane unreadable. Never had these issues in
Kitty.

Hopefully these issues are fixed when Wezterm gets updated.