Disable the process spinner
This process spinner has been the cause of so much flickering of the
cursor in the terminals that I've used. I think only Kitty manages to
not flicker the cursor when the process spinner is active. I don't know
how Kitty manages to not flicker the cursor, most probably because it's
rendering speeds are just insane, or it does some special optimisation,
but it's great.

Ghostty and Wezterm will flicker the cursor, sometimes after a while
when typing the commit message, and sometimes immediately. It makes for
a terrible experience writing commit messages. Wezterm's flicker is
quite a bit worse than Ghostty's flicker, and happens more often as
well.