Disable the resize overlay
It is more annoying than helpful, especially since I'm using a tiling
window manager that automatically does the resizing for me.

It makes sense for manually resizing a window in a desktop environment,
but I'm not using a desktop environment and my window manager handles
all the resizing for me.