Remove the keymap to move to the beginning of the line using ^
Yazi does not have a replace mode and hence can't replace characters.
This makes this keymap useless, as you would only use ^ if you want to
stay in normal mode but move to the beginning of the line. Hence, you
would just use I to insert text at the beginning of the line in Yazi.