Added more language servers to Emacs
Specifically, basedpyright and efm language server, while also migrating
from LTEX to LTEX+.

Also made use of LSP UI functions instead of the default LSP functions
to match the Neovim configuration more closely. The documentation
position has also been changed to be at the cursor instead of at the
bottom of the screen, matching the Neovim configuration.

Changed all the verbatim words to words, except for the one for the
file path, as the file path contains a ~ character.

Added all the language servers and the formatter used to the external
programs section. Also added Jupyter Notebook to the same section as it
is needed for the ipynb exporter to automatically open exported files.