Migrate to basedpyright instead of pylsp
Pylsp has been extremely slow for some reason, causing me to constantly
restart Neovim to get it to respond. As such, I'm migrating to
basedpyright as it is easier to install on my system and can be used
with my Emacs configuration.

Also migrated from Black to Ruff as a formatter, since I used to use
Black for formatting with Pylsp, but without Pylsp, installing the isort
extension isn't that simple, so I'm going to just use Ruff which is way
way faster, but I'll have to add comment blocks add the start of each
line to get the spacing I want, just like with most other programming
languages.

LTEX has also been replaced with LTEX+, and all the shared language
servers with Emacs are set up separately from mason-lspconfig,
since they are not installed by Mason.

Also added selene as a Lua linter, though so far it's more annoying than
anything else.