Removed Aerial.nvim in favour of treesitter context
I don't use Aerial.nvim for anything other than for the win bar, where
it shows the current function after the file name. Treesitter context
is much better as I can see the line number where the function is on,
and it also works on things like table keys instead of just functions,
which makes it much more useful. Super easy to jump to the start of the
context if I need to do so, which I can't really do in Aerial without
opening the Aerial menu.

I also much rather look for my document symbols using the fuzzy finder
instead of going through a menu to search for it manually.