Add precognition and nvim-training
Also renamed a few files for better readability and moved some files
around to make it easier to find the configurations for the plugins.

Moved all game like plugins into a games.lua file.

The init.lua file now returns an empty list of plugins, so that
lazy.nvim can discover the other plugins in the directory, as all the
plugins are categorised in their own files.