Remove precognition.nvim
It doesn't do anything for me, since it only hints the most obvious
jumps with b, e, w, 0, ^ and $. There's not f, F, t or T jumps, which
was what I was hoping for, or some multiple of b, e, or w to get to a
specific character.

I'm basically already using all of its hints, maybe save for the curly
braces and the 0.

Worst of all, it keeps shifting lines around, making me think that there
are lines when there isn't actually a line, and just making coding
really confusing for me.

Its hints are wrong half the time as well, such as pressing e to move
to a word that is behind the current cursor position, or w to move to
the current cursor position, which just doesn't make any sense.