Switch back to kitty
Ghostty was kinda nice, but there were some issues that ultimately made
me switch back to Kitty. One of the more annoying ones was the image
preview in Yazi not being displayed fully, always cutting off the
right side of the image. This is fine most of the times, but when there
are important details at that area, it forces me to open the image to
have a better look at it, which is quite frustrating. Ghostty is also
quite noisy, but it was alright for me, though it was quite annoying to
see the notification pop up when I copied the file name in Yazi when I
wasn't expecting it.

Furthermore, since Kitty supports changing themes based on the current
mode, I don't really have much of a reason to use Ghostty anymore. Kitty
also makes far more efficient use of space with the new Maple Mono NF CN
font that I'm using compared to Ghostty. I don't even have to configure
anything and Kitty just makes sure there is no random weird spacing
anywhere.

The killer feature that Kitty has that none of the other terminal
emulators have is the layout feature, it's just absolutely amazing how I
don't have to think about opening new splits and can just open one
knowing it would be in the right spot. Kitty's launch with the current
directory also works with Yazi, which means running things that require
some output has never been easier, and Ghostty can't do it.

This is not at all mentioning how Kitty is still the fastest terminal
around despite being written in Python, with image previews being faster
than Ghostty, and how it doesn't cause insane cursor flickering with the
Neogit spinner when all other terminals fall victim to it.

Also added a bunch of key binds to Wezterm and Kitty to follow the key
binds used in Ghostty.