Switch back to ghostty
For some reason Kitty has some weird rendering issue when using Yazi.
The git icons become much smaller after opening a file that isn't the
file with the git icon, and then it gets bigger again after you hover
over the file. Weird.

It also turns out that the padding is the cause of the image being cut
off, so removing the left and right padding fixes that issue.