Switch back to Kitty
The image being cut off in Ghostty is still there, it's just not as
noticeable. It only becomes noticeable in split windows since I set the
padding to zero, but it is still really annoying.

I also really miss opening tabs in the current working directory and the
grid layout.

I figured that I can fix the Kitty issue by just setting a colour for
the Git icons, so I'm just going to do that. It's even documented in
Yazi's FAQ, linked below.

https://yazi-rs.github.io/docs/faq#why-icons-shrink