Disable linewise hybrid mode
The layout shift when making use of lists in markdown with this is just
unbearable to look at and makes things feel so slow.

I will have to get used to using Ctrl + O to temporarily enter normal
mode to move in insert mode though, since the usual method of just
leaving insert mode will cause the layout to shift, which is super
annoying when editing.